rename api files to follow url names
