+++ src/api/routes/admin/index.js
@@ -27,13 +27,13 @@ export default (app, rootDirectory) => {
   route.get(
     "/fulfillments/:fulfillment_id/shippo/order",
     authenticate(),
-    middlewares.wrap(require("./shippo-order").default)
+    middlewares.wrap(require("./fulfillment-order").default)
   )
 
   route.get(
     "/fulfillments/:fulfillment_id/shippo/packingslip",
     authenticate(),
-    middlewares.wrap(require("./order-packingslip").default)
+    middlewares.wrap(require("./fulfillment-packingslip").default)
   )
 
   // all your errors are belong to this
